Fixed linkage errors. Size_t had to be changed to int because python does not support unsigend ints and caused problems. In order to fix the linkage issue I added spkmeans.c file into the sources list in the extension (at setup.py)

+++ spkmeans.h
@@ -1,12 +1,12 @@
 #ifndef SPKMEANS_H_
 #define SPKMEANS_H_
 double **kmeans(double ** centeroids, double ** vectors, int const length, int const numberOfVectors, int const K, double const EPSILON, int const MAX_ITER);
-double **wam(double **matrix, size_t numberOfVectors, size_t length);
-double ** ddg(double **matrix, size_t numberOfVectors, size_t length);
-double **lnorm(double ** vectors, size_t numberOfVectors, size_t length);
-double ** jacobi(double **vectors, size_t dim);
-double **createBlockMatrix(size_t size, size_t numberOfRows, size_t numberOfColumns);
+double **wam(double **matrix, int numberOfVectors, int length);
+double ** ddg(double **matrix, int numberOfVectors, int length);
+double **lnorm(double ** vectors, int numberOfVectors, int length);
+double ** jacobi(double **vectors, int dim);
+double **createBlockMatrix(int size, int numberOfRows, int numberOfColumns);
 void freeBlock(double **ptr);
 void errorMsg(int code);
-void printMatrix(double **array, size_t numberOfRows, size_t numberOfColumns);
+void printMatrix(double **array, int numberOfRows, int numberOfColumns);
 #endif
